jumper/spawncooldown/firstkillshield fix

+++ TheEpicRoles/Main.cs
@@ -21,7 +21,7 @@ namespace TheEpicRoles {
     {
         public const string Id = "me.laicosvk.theepicroles";
         public const string VersionString = "1.0.4";
-        public static byte firstKill = 0;
+        public static uint firstKill = 0;
 
         public static System.Version Version = System.Version.Parse(VersionString);
